refactor: use reusable ui component
